Relocated firebase.config's expected location to the root and fixed 'textAnchor' className
