+++ src/db/user-collection.js
@@ -8,7 +8,7 @@ import {
     updateDoc,
     where,
   } from "firebase/firestore";
-import { db } from "../firebase.config";
+import { db } from "../../firebase.config";
 
 const usersRef = collection(db, "users");
